# contracts/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Platform V2 API",
...
    "version": "0.5.0"
  },
  "paths": {
...
    "/v1/integrations/aws": {
      "post": {
        "tags": [
          "integrations"
        ],
        "summary": "Register AWS integration",
        "description": "Connect a customer AWS account via cross-account IAM role + external ID. Requires **tenant_admin** or **super_admin**.\n\nCustomer must trust the hub account and grant read-only permissions. `external_id` is stored encrypted and never returned in responses.\n\n- **409** \u2014 integration already exists for this account\n- **400** \u2014 validation error (account_id, regions, external_id length)",
        "operationId": "register_aws_integration_v1_integrations_aws_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/AwsIntegrationCreate"
              }
            }
          },
          "required": true
        },
        "responses": {
          "201": {
            "description": "Integration created",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/IntegrationResponse"
                }
              }
            }
          },
          "409": {
            "description": "Duplicate account"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/integrations": {
      "get": {
        "tags": [
          "integrations"
        ],
        "summary": "List AWS integrations",
        "description": "All integrations for the authenticated tenant. Use `id` when calling **POST /v1/scans**.",
        "operationId": "list_integrations_v1_integrations_get",
        "responses": {
          "200": {
            "description": "Integration list (may be empty)",
            "content": {
              "application/json": {
                "schema": {
                  "items": {
                    "$ref": "#/components/schemas/IntegrationResponse"
                  },
                  "type": "array",
                  "title": "Response List Integrations V1 Integrations Get"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/findings": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "List findings for scan",
        "description": "Policy evaluation results for a scan. Filter by `result=fail` for issues only. Includes remediation copy from policy YAML catalog.",
        "operationId": "list_findings_v1_findings_get",
        "parameters": [
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Scan UUID",
              "title": "Scan Id"
            },
            "description": "Scan UUID"
          },
          {
            "name": "result",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "pass | fail | error",
              "title": "Result"
            },
            "description": "pass | fail | error"
          },
          {
            "name": "policy_id",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "Filter e.g. AWS_S3_001",
              "title": "Policy Id"
            },
            "description": "Filter e.g. AWS_S3_001"
          },
          {
            "name": "status",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "open | resolved | suppressed",
              "title": "Status"
            },
            "description": "open | resolved | suppressed"
          },
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "maximum": 500,
              "minimum": 1,
              "default": 100,
              "title": "Limit"
            }
          },
          {
            "name": "offset",
            "in": "query",
            "required": false,
            "schema": {
              "type": "integer",
              "minimum": 0,
              "default": 0,
              "title": "Offset"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Finding list",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/FindingResponse"
                  },
                  "title": "Response List Findings V1 Findings Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
    "/v1/findings/{finding_id}": {
      "get": {
        "tags": [
          "findings"
        ],
        "summary": "Get finding by ID",
        "description": "Single finding with full remediation block. Requires `scan_id` query param for tenant scoping.",
        "operationId": "get_finding_v1_findings__finding_id__get",
        "parameters": [
          {
            "name": "finding_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Finding Id"
            }
          },
          {
            "name": "scan_id",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "description": "Scan UUID",
              "title": "Scan Id"
            },
            "description": "Scan UUID"
          }
        ],
        "responses": {
          "200": {
            "description": "Finding detail",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/FindingResponse"
                }
              }
            }
          },
          "404": {
            "description": "Finding not found"
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
    "/v1/admin/tenants/{tenant_id}/integrations": {
      "get": {
        "tags": [
          "admin"
        ],
        "summary": "List tenant integrations (admin)",
        "description": "Same as tenant-scoped integrations list but callable by super_admin for any tenant.",
        "operationId": "list_tenant_integrations_v1_admin_tenants__tenant_id__integrations_get",
        "parameters": [
          {
            "name": "tenant_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid",
              "title": "Tenant Id"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": {
                    "$ref": "#/components/schemas/AdminIntegrationResponse"
                  },
                  "title": "Response List Tenant Integrations V1 Admin Tenants  Tenant Id  Integrations Get"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        },
        "security": [
          {
            "SupabaseJWT": []
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "AdminIntegrationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "role_arn": {
            "type": "string",
            "title": "Role Arn"
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Regions"
          },
          "status": {
            "type": "string",
            "title": "Status"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "provider",
          "account_id",
          "role_arn",
          "regions",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "AdminIntegrationResponse"
      },
...
      "AwsIntegrationCreate": {
        "properties": {
          "account_id": {
            "type": "string",
            "maxLength": 12,
            "minLength": 12,
            "title": "Account Id",
            "description": "12-digit AWS account ID",
            "examples": [
              "387957186076"
            ]
          },
          "role_arn": {
            "type": "string",
            "minLength": 20,
            "title": "Role Arn",
            "description": "Customer IAM role ARN the hub account can AssumeRole into",
            "examples": [
              "arn:aws:iam::387957186076:role/SteampipeReadRole"
            ]
          },
          "external_id": {
            "type": "string",
            "minLength": 8,
            "title": "External Id",
            "description": "STS external ID (encrypted at rest, min 8 chars)"
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "minItems": 1,
            "title": "Regions",
            "description": "AWS regions to scan",
            "examples": [
              [
                "us-east-1"
              ]
            ]
          }
        },
        "type": "object",
        "required": [
          "account_id",
          "role_arn",
          "external_id",
          "regions"
        ],
        "title": "AwsIntegrationCreate"
      },
...
      "FindingResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "policy_id": {
            "type": "string",
            "title": "Policy Id",
            "description": "Unified policy ID e.g. AWS_S3_001"
          },
          "resource_id": {
            "type": "string",
            "title": "Resource Id"
          },
          "resource_type": {
            "type": "string",
            "title": "Resource Type"
          },
          "result": {
            "type": "string",
            "title": "Result",
            "description": "pass | fail | error"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "open | resolved | suppressed"
          },
          "severity": {
            "type": "string",
            "title": "Severity"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "description": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Description"
          },
          "evidence": {
            "additionalProperties": true,
            "type": "object",
            "title": "Evidence",
            "description": "Fields that triggered the policy"
          },
          "remediation": {
            "$ref": "#/components/schemas/RemediationResponse"
          },
          "evaluated_at": {
            "type": "string",
            "title": "Evaluated At"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "policy_id",
          "resource_id",
          "resource_type",
          "result",
          "status",
          "severity",
          "title",
          "evidence",
          "remediation",
          "evaluated_at",
          "created_at"
        ],
        "title": "FindingResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
      "IntegrationResponse": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "tenant_id": {
            "type": "string",
            "title": "Tenant Id"
          },
          "provider": {
            "type": "string",
            "title": "Provider"
          },
          "account_id": {
            "type": "string",
            "title": "Account Id"
          },
          "role_arn": {
            "type": "string",
            "title": "Role Arn"
          },
          "regions": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Regions"
          },
          "status": {
            "type": "string",
            "title": "Status",
            "description": "active | invalid | disabled"
          },
          "created_at": {
            "type": "string",
            "title": "Created At"
          },
          "updated_at": {
            "type": "string",
            "title": "Updated At"
          }
        },
        "type": "object",
        "required": [
          "id",
          "tenant_id",
          "provider",
          "account_id",
          "role_arn",
          "regions",
          "status",
          "created_at",
          "updated_at"
        ],
        "title": "IntegrationResponse"
      },
...
      "RemediationResponse": {
        "properties": {
          "headline": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Headline",
            "description": "Short risk headline for UI"
          },
          "risk_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Risk Summary"
          },
          "business_impact": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Business Impact"
          },
          "fix_summary": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Fix Summary"
          },
          "estimated_fix_minutes": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Estimated Fix Minutes"
          },
          "framework_mappings": {
            "items": {
              "type": "string"
            },
            "type": "array",
            "title": "Framework Mappings",
            "description": "CIS / SOC2 control refs"
          },
          "aws_cli": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Aws Cli",
            "description": "Suggested AWS CLI remediation"
          },
          "terraform": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Terraform"
          },
          "cloudformation": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Cloudformation"
          }
        },
        "type": "object",
        "title": "RemediationResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    },
    "securitySchemes": {
      "SupabaseJWT": {
        "type": "http",
        "scheme": "bearer",
        "bearerFormat": "JWT",
        "description": "Supabase `access_token` from POST .../auth/v1/token?grant_type=password. Paste token only in Authorize dialog."
      },
      "DevTenantId": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Tenant-ID",
        "description": "Tenant UUID. Only when DEV_TENANT_HEADER_AUTH=true."
      },
      "DevRole": {
        "type": "apiKey",
        "in": "header",
        "name": "X-Role",
        "description": "tenant_admin | viewer | super_admin. Pair with DevTenantId."
      }
    }
  },
...
  "servers": [
    {
      "url": "http://localhost:8090",
      "description": "Local dev"
    },
    {
      "url": "https://api.example.com",
      "description": "Staging / production"
    }
  ]
}